+++ Cargo.toml
@@ -23,3 +23,5 @@ tokenizers = "0.13.3"
 tokio = { version = "1.27.0", features = ["macros", "io-util", "rt", "rt-multi-thread"] }
 tiktoken-rs = { version = "0.4.1", features = ["async-openai"] }
 mlua = { version = "0.8.8", features = [ "lua54", "vendored", "async" ] }
+llama-rs = { git = "https://github.com/rustformers/llama-rs" }
+rand = "0.8.5"
